size

+++ 0x00-hello_world/6-size.c
@@ -5,9 +5,9 @@
  */
 int main(void)
 {
-printf("Size of a char: %lu byte(s)", sizeof(char));
-printf("Size of an int: %lu byte(s)", sizeof(int));
-printf("Size of a long int: %lu byte(s)", sizeof(long int));
-printf("Size of a long long int: %lu byte(s)", sizeof(long long int));
-printf("Size of a float: %lu byte(s)", sizeof(float));
+printf("Size of a char: %lu byte(s)", (unsigned long)sizeof(char));
+printf("Size of an int: %lu byte(s)", (unsigned long)sizeof(int));
+printf("Size of a long int: %lu byte(s)", (unsigned long)sizeof(long int));
+printf("Size of a long long int: %lu byte(s)", (unsigned long)sizeof(long long int));
+printf("Size of a float: %lu byte(s)", (unsigned long)sizeof(float));
 return (0);
